fix: #44 pr 수정사항 적용
